moveonly: move WalletStorage to separate file
