+++ src/wallet/storage.h
@@ -1,0 +1,37 @@
+// Copyright (c) 2019-2020 The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BITCOIN_WALLET_STORAGE_H
+#define BITCOIN_WALLET_STORAGE_H
+
+#include <wallet/crypter.h>
+#include <wallet/db.h>
+#include <wallet/walletdb.h>
+#include <wallet/walletutil.h>
+
+#include <string>
+
+namespace wallet {
+// Wallet storage things that ScriptPubKeyMans need in order to be able to store things to the wallet database.
+// It provides access to things that are part of the entire wallet and not specific to a ScriptPubKeyMan such as
+// wallet flags, wallet version, encryption keys, encryption status, and the database itself. This allows a
+// ScriptPubKeyMan to have callbacks into CWallet without causing a circular dependency.
+// WalletStorage should be the same for all ScriptPubKeyMans of a wallet.
+class WalletStorage
+{
+public:
+    virtual ~WalletStorage() = default;
+    virtual const std::string GetDisplayName() const = 0;
+    virtual WalletDatabase& GetDatabase() const = 0;
+    virtual bool IsWalletFlagSet(uint64_t) const = 0;
+    virtual void UnsetBlankWalletFlag(WalletBatch&) = 0;
+    virtual bool CanSupportFeature(enum WalletFeature) const = 0;
+    virtual void SetMinVersion(enum WalletFeature, WalletBatch* = nullptr) = 0;
+    virtual const CKeyingMaterial& GetEncryptionKey() const = 0;
+    virtual bool HasEncryptionKeys() const = 0;
+    virtual bool IsLocked() const = 0;
+};
+} // namespace wallet
+
+#endif // BITCOIN_WALLET_STORAGE_H
